+++ src/Principal.java
@@ -13,10 +13,10 @@ public class Principal {
             arqCurso = new ArquivoCurso();
 
             ControleUsuario ctrlUsuario = new ControleUsuario(arqUsuario, arqCurso);
-            ControleCurso ctrlCurso = new ControleCurso(arqCurso);
+            ControleCurso ctrlCurso = new ControleCurso(arqCurso, arqUsuario);
 
             VisaoCurso visaoCurso = new VisaoCurso(ctrlCurso, console);
-            VisaoUsuario visaoUsuario = new VisaoUsuario(ctrlUsuario, visaoCurso,  ctrlCurso, console);
+            VisaoUsuario visaoUsuario = new VisaoUsuario(ctrlUsuario, visaoCurso, console);
 
             visaoUsuario.menuUsuario();